+++ packages/cta-link/src/js/react.js
@@ -21,17 +21,28 @@ import PropTypes from 'prop-types';
  * DEFAULT
  * The cta link component
  *
+ * @param  {string}  linkComponent    - The component used for the link
  * @param  {boolean} dark             - Add the dark variation class, optional
  * @param  {string}  link             - The link target, if not supplied this will render as a button, optional
  * @param  {string}  text             - The text of the CTA link
  * @param  {string}  className        - An additional class, optional
  * @param  {object}  attributeOptions - Any other attribute options
  */
-const AUctaLink = ({ dark, link, text, className = '', ...attributeOptions }) => {
+const AUctaLink = ({ linkComponent, dark, link, text, className = '', ...attributeOptions }) => {
+	const LinkComponent = linkComponent;
 
 	if( link ) {
+		// If we are using a normal link
+		if( LinkComponent === 'a' ) {
+			attributeOptions.href = link;
+		}
+		// If we are using a link component
+		else if( typeof LinkComponent === 'function' ) {
+			attributeOptions.to = link;
+		}
+
 		return (
-			<a className={ `au-cta-link ${ className }${ dark ? ` au-cta-link--dark` : '' }` } href={ link } { ...attributeOptions }>{ text }</a>
+			<LinkComponent className={ `au-cta-link ${ className }${ dark ? ` au-cta-link--dark` : '' }` } { ...attributeOptions }>{ text }</LinkComponent>
 		);
 	}
 	else {
@@ -46,6 +57,11 @@ AUctaLink.propTypes = {
 	link: PropTypes.string,
 	text: PropTypes.string.isRequired,
 	className: PropTypes.string,
+	linkComponent: PropTypes.oneOfType([ PropTypes.string, PropTypes.func ])
+};
+
+AUctaLink.defaultProps = {
+	linkComponent: 'a',
 };
 
 export default AUctaLink;