+++ packages/link-list/src/js/react.js
@@ -19,13 +19,16 @@ import PropTypes from 'prop-types';
 /**
  * An item inside the AUlinkList component
  *
+ * @param  {node}   text             - The link Text or link html
  * @param  {string} link             - The link URL, optional
- * @param  {node} text               - The link Text or link html
+ * @param  {string} linkComponent    - The component used for the link
  * @param  {object} li               - An additional object to be spread into the wrapping element, optional
  * @param  {object} attributeOptions - Any other attribute options, optional
  */
-export const AUlinkListItem = ({ text, link, li = {}, ...attributeOptions }) => {
+export const AUlinkListItem = ({ text, link, linkComponent, li = {}, children, ...attributeOptions }) => {
+	const LinkComponent = linkComponent;
 
+	// If there is no link provided and an onClick function
 	if( typeof onClick === 'function' ) {
 		attributeOptions.onClick = onClick;
 
@@ -35,21 +38,35 @@ export const AUlinkListItem = ({ text, link, li = {}, ...attributeOptions }) => 
 		}
 	}
 
-	return (
-		<li { ...li }>
-			{
-				link === undefined
-					? text
-					: <a href={ link } { ...attributeOptions }>{ text }</a>
-			}
-		</li>
-	);
+	// If we are using a normal link
+	if( LinkComponent === 'a' ) {
+		attributeOptions.href = link;
+	}
+	// If we are using a link component
+	else if( typeof LinkComponent === 'function' ) {
+		attributeOptions.to = link;
+	}
+
+	if( link ){
+		return (
+			<li { ...li }>
+				<LinkComponent { ...attributeOptions }>{ text }{ children }</LinkComponent>
+			</li>
+		)
+	}
+
+	return ( <li { ...li }>{ text }{ children }</li> );
 };
 
 AUlinkListItem.propTypes = {
-	text: PropTypes.string.isRequired,
+	text: PropTypes.node.isRequired,
 	link: PropTypes.string,
 	li: PropTypes.object,
+	linkComponent: PropTypes.oneOfType([ PropTypes.string, PropTypes.func ])
+};
+
+AUlinkListItem.defaultProps = {
+	linkComponent: "a",
 };
 
 
@@ -59,13 +76,14 @@ AUlinkListItem.propTypes = {
  *
  * @param  {array}   items            - All items inside the link list to be passed to AUlinkListItem, format: { link: '', text: '', onClick: () }
  * @param  {string}  className        - An additional class, optional
+ * @param  {string}  linkComponent    - The component used for the link
  * @param  {object}  attributeOptions - Any other attribute options, optional
  */
-const AUlinkList = ({ inline, items, className = '', ...attributeOptions }) => (
+const AUlinkList = ({ inline, items, linkComponent, className = '', ...attributeOptions }) => (
 	<ul className={ `au-link-list ${ className }${ inline ? ' au-link-list--inline' : '' }` } { ...attributeOptions }>
 		{
 			items.map(
-				( item, i ) => <AUlinkListItem key={ i } { ...item } />
+				( item, i ) => <AUlinkListItem linkComponent={ linkComponent } key={ i } { ...item } />
 			)
 		}
 	</ul>
@@ -80,6 +98,12 @@ AUlinkList.propTypes = {
 			li: PropTypes.object,
 		})
 	).isRequired,
+	linkComponent: PropTypes.oneOfType([ PropTypes.string, PropTypes.func ])
 };
+
+AUlinkList.defaultProps = {
+	linkComponent: "a",
+};
+
 
 export default AUlinkList;