+++ packages/body/CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ## Versions
 
+* [v2.0.6 - Change npm run watch browser-sync location](#v206)
 * [v2.0.5 - Update dependencies](#v205)
 * [v2.0.4 - Fix typo in homepage link](#v204)
 * [v2.0.3 - Changing homepage link](#v203)
@@ -31,6 +32,11 @@
 
 
 ## Release History
+
+### v2.0.6
+
+- Change npm run watch browser-sync location
+
 
 ### v2.0.5