+++ packages/core/CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ## Versions
 
+* [v2.0.3 - Change npm run watch browser-sync location](#v203)
 * [v2.0.2 - Update dependencies](#v202)
 * [v2.0.1 - Change homepage link](#v201)
 * [v2.0.0 - Changing mix values for border and muted, focus colour, adding offset to focus](#v200)
@@ -30,6 +31,11 @@
 
 
 ## Release History
+
+### v2.0.3
+
+- Change npm run watch browser-sync location
+
 
 ### v2.0.2