+++ packages/text-inputs/CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ## Versions
 
+* [v2.0.5 - Change npm run watch browser-sync location](#v205)
 * [v2.0.4 - Update dependencies](#v204)
 * [v2.0.3 - Change homepage links](#v203)
 * [v2.0.2 - Adding focus offset back to text inputs](#v202)
@@ -29,6 +30,11 @@
 
 
 ## Release History
+
+### v2.0.5
+
+- Change npm run watch browser-sync location
+
 
 ### v2.0.4